Arrumar user

+++ Controller/User.controller.php
@@ -1,10 +1,10 @@
 <?php
 require_once __DIR__ .'/../Utils/autoload.php';
 include_once __DIR__ . '/../Model/User.class.php';
+include_once __DIR__ . '/../Model/Bike.class.php';
+include_once __DIR__ . '/../Model/Bicicletario.class.php';
 
 $acao = $_GET['acao'];
-
-$cor = $_POST['cor'];
 
 
 if ($acao == 'cadastrar'){
@@ -13,6 +13,8 @@ if ($acao == 'cadastrar'){
     $user->setNome($_POST['nome']);
     
     $user->save();
+
+    $cor = $_POST['cor'];
 
     $bike = new Bike();
     $bike->setCor($_POST['cor']);
@@ -24,12 +26,21 @@ if ($acao == 'cadastrar'){
     $imagePath = $_FILES['imagem']['tmp_name'];
 
     move_uploaded_file($imagePath,$savePath);
-    header('Location:../View/Func/Gere.User.php');
+    // header('Location:../View/Func/Gere.User.php');
 }
 
 else if($acao == 'deletar'){
-    Bike::deleteUser($_REQUEST['id']);
-    User::delete($_REQUEST['id']);    
+    $temp = Bicicletario::loadByBike($_REQUEST['id']);
+    if($temp == false){
+        Bike::deleteUser($_REQUEST['id']);
+        User::delete($_REQUEST['id']); 
+    }else{
+        $message = new Message();
+        $message->setTipo("danger");
+        $message->setConteudo("Usuario cadastrado no Bicicletario");
+
+        header('Location:../View/Func/Gere.Bike.php?message=' . $message->__toString());
+    }   
     header('Location:../View/Func/Gere.User.php');
     
 } else if($acao == 'atualizar'){
